+++ src/hooks/index.ts
@@ -16,8 +16,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import useFeedbackMessage from './useFeedbackMessage'
-import useTheme from './useTheme'
+import { useFeedbackMessage } from './useFeedbackMessage'
+import { useTheme } from './useTheme'
 
 export default {
   useFeedbackMessage,
